Missing fields when model is ref'd more than once
I had the situation where I had the following models:

```
class Base(Model):
    pass

class Sub(Model):
    base = ForeignKey("Base", ...)

class Comment(Model):
    text = TextField()
    base = ForeignKey("Base", related_name="comments", ...)
    sub = ForeignKey("Base", related_name="comments", null=True, ...)

class OurQLSchema(DjangoQLSchema):
    def get_fields(self, model):
        if model == Base:
            return ['comments']
        if model == Sub:
            return ['comments']
        return super(UserQLSchema, self).get_fields(model)
```

In the above example, the query "sub.comments.text" would not work. This is
because by the time the schema would try to inspect "sub.comments", it would
already have seen the Comment model, would not add the field. I believe this
to be incorrect. While the comment model does not need to be reinspected, the
field still needs to be added.